Uses standard M-Lab filesystem paths for experiments
Currently, the directory structure of Wehe does not align well with the
standard filesystem layout for M-lab experiments. M-Lab is, where
possible, migrating containers to run as non-root users. The current
Wehe directory layout conflicts with this, as it wants to write to
directories that are owned by root. The changes in the commit bring the
directory structure of Wehe output in line with M-Lab standard layouts.